+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/simulation/SemThenDiscretizeSimulation.java
@@ -2,26 +2,26 @@ package edu.cmu.tetrad.algcomparison.simulation;
 
 import edu.cmu.tetrad.data.DataType;
 import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.data.Discretizer;
 import edu.cmu.tetrad.graph.Graph;
 import edu.cmu.tetrad.graph.GraphUtils;
+import edu.cmu.tetrad.graph.Node;
 import edu.cmu.tetrad.sem.SemIm;
-import edu.cmu.tetrad.sem.SemImInitializationParams;
 import edu.cmu.tetrad.sem.SemPm;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 /**
  * @author jdramsey
  */
-public class ContinuousLinearGaussianSemSimulation implements Simulation {
+public class SemThenDiscretizeSimulation implements Simulation {
     private Graph graph;
     private List<DataSet> dataSets;
 
     @Override
     public void createData(Parameters parameters) {
-        dataSets = new ArrayList<>();
-
         int numEdges = parameters.getInt("numEdges");
 
         if (numEdges == -1) {
@@ -37,18 +37,11 @@ public class ContinuousLinearGaussianSemSimulation implements Simulation {
                 parameters.getInt("maxOutdegree"),
                 parameters.getInt("connected") == 1);
 
-        for (int i = 0; i < parameters.getInt("numRuns"); i++) {
-            SemPm pm = new SemPm(graph);
-            SemImInitializationParams params = new SemImInitializationParams();
-            params.setVarRange(parameters.getDouble("variance"), parameters.getDouble("variance"));
-            SemIm im = new SemIm(pm);
-            dataSets.add(im.simulateData(parameters.getInt("sampleSize"), false));
-        }
-    }
+        dataSets = new ArrayList<>();
 
-    @Override
-    public DataSet getDataSet(int index) {
-        return dataSets.get(index);
+        for (int i = 0; i < parameters.getInt("numRuns"); i++) {
+            dataSets.add(simulate(graph, parameters));
+        }
     }
 
     @Override
@@ -58,7 +51,7 @@ public class ContinuousLinearGaussianSemSimulation implements Simulation {
 
     @Override
     public String getDescription() {
-        return "Linear, Gaussian SEM simulation";
+        return "Simulation SEM data then discretizing some variables";
     }
 
     @Override
@@ -74,6 +67,7 @@ public class ContinuousLinearGaussianSemSimulation implements Simulation {
         parameters.add("numRuns");
         parameters.add("sampleSize");
         parameters.add("variance");
+        parameters.add("numCategories");
         return parameters;
     }
 
@@ -83,7 +77,29 @@ public class ContinuousLinearGaussianSemSimulation implements Simulation {
     }
 
     @Override
+    public DataSet getDataSet(int index) {
+        return dataSets.get(index);
+    }
+
+    @Override
     public DataType getDataType() {
-        return DataType.Continuous;
+        return DataType.Mixed;
+    }
+
+    private DataSet simulate(Graph graph, Parameters parameters) {
+        SemPm pm = new SemPm(graph);
+        SemIm im = new SemIm(pm);
+        DataSet continuousData = im.simulateData(parameters.getInt("sampleSize"), false);
+
+        List<Node> shuffledNodes = new ArrayList<>(continuousData.getVariables());
+        Collections.shuffle(shuffledNodes);
+
+        Discretizer discretizer = new Discretizer(continuousData);
+
+        for (int i = 0; i < shuffledNodes.size() * parameters.getDouble("percentDiscreteForMixedSimulation") * 0.01; i++) {
+            discretizer.equalIntervals(shuffledNodes.get(i), parameters.getInt("numCategories"));
+        }
+
+        return discretizer.discretize();
     }
 }
